Windows client port, protocol freeze, wire-compatible hardening
Windows (MinGW-w64) support:
- port.h/port.c: platform layer (winsock2 wrappers with errno mapping,
  WSAPoll, WSABUF batch emulation for recvmmsg/sendmmsg, UDP GSO via
  SIO_UDP_NETSEGMENT with fallback, eventfd via loopback UDP socketpair,
  QPC clock, BCryptGenRandom, CreateProcess helpers, ctrl handler)
- tun_win.c: wintun backend (runtime LoadLibrary, single-session pool
  reader); tun.c stays Linux-only
- route.c: netsh/iphlpapi Windows backend mirroring the Linux sequence
- https.c: in-process libssl transport replacing fork+openssl s_client
  (system ROOT store on Windows, SSL_CERT_FILE/bundle paths on Linux)
- netstack/socks/proxy datapath via port_* wrappers; Linux path is
  byte-identical inline passthroughs
- Makefile TARGET=win32 + CI win-cross job (mingw cross, PE32+ check,
  wine --help + real wire smoke vs Linux iwan-server --no-tun)
- iwan-server stays Linux-only; socks/ping/auth/OIDC need no admin

Protocol freeze:
- PROTOCOL.md: byte-exact wire spec + frozen security model
- protocol.h: IWAN_PROTO_VERSION (informational) + _Static_asserts
  pinning layout constants; verified live against the reference wire

Wire-compatible hardening (no byte changes):
- server: per-source token-mismatch rate limit (4/s) on DATA/CLOSE,
  constant-time token compare, rebind gated on clean mismatch history
- client: OPEN_REJECT signature check; fail-closed RNG (no weak
  fallback); gateway-vs-server mismatch warning

Verified: Linux build clean, win32 cross build clean (0 warnings),
Linux client and Windows client (wine) both complete PING and
OPEN/OPEN_ACK/CLOSE against a Linux iwan-server; OIDC TLS flow reaches
the live controller from both platforms.

+++ src/common/json.c
@@ -52,8 +52,9 @@ static void p_fail(struct P *p, size_t off, int eof, const char *msg)
             last_nl = k;
         }
     col = last_nl == (size_t)-1 ? i : i - last_nl - 1;
-    snprintf(p->err, sizeof p->err, "%s at line %zu column %zu", msg, line,
-             col);
+    /* %llu not %zu: msvcrt printf (Windows) lacks %zu */
+    snprintf(p->err, sizeof p->err, "%s at line %llu column %llu", msg,
+             (unsigned long long)line, (unsigned long long)col);
 }
 
 struct jbuf {
